% Auto-generated: do not edit by hand
\name{dashQueryBuilder}

\alias{dashQueryBuilder}

\title{DashQueryBuilder component}

\description{
Component description
}

\usage{
dashQueryBuilder(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

