% Auto-generated: do not edit by hand
\name{dqbDashQueryBuilder}

\alias{dqbDashQueryBuilder}

\title{DashQueryBuilder component}

\description{

}

\usage{
dqbDashQueryBuilder(id=NULL, fields=NULL, theme=NULL, tree=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{fields}{Logical | numeric | character | named list | unnamed list. }

\item{theme}{Character. }

\item{tree}{Logical | numeric | character | named list | unnamed list. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

