% Auto-generated: do not edit by hand
\name{dqbDashQueryBuilder}

\alias{dqbDashQueryBuilder}

\title{DashQueryBuilder component}

\description{

}

\usage{
dqbDashQueryBuilder(id=NULL, fields=NULL, jsonLogicFormat=NULL,
mongodbFormat=NULL, queryStringFormat=NULL, sqlFormat=NULL,
theme=NULL, tree=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{fields}{Logical | numeric | character | named list | unnamed list. }

\item{jsonLogicFormat}{Character. }

\item{mongodbFormat}{Character. }

\item{queryStringFormat}{Character. }

\item{sqlFormat}{Character. }

\item{theme}{A value equal to: 'material', 'antd', 'basic'. }

\item{tree}{Logical | numeric | character | named list | unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

